requirements.txt
blis==0.7.7
catalogue==2.0.7
certifi==2022.5.18
charset-normalizer==2.0.12
click==8.1.3
cymem==2.0.6
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.3.0/en_core_web_md-3.3.0-py3-none-any.whl
greenlet==1.1.2
idna==3.3
IMDbPY==2021.4.18
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
lxml==4.8.0
MarkupSafe==2.1.1
murmurhash==1.0.7
nltk==3.7
numpy==1.22.3
packaging==21.3
pathy==0.6.1
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.9
regex==2022.4.24
requests==2.27.1
smart-open==5.2.1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
SQLAlchemy==1.4.36
srsly==2.4.3
thinc==8.0.15
tqdm==4.64.0
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
